+++ hutool-core/src/main/java/cn/hutool/core/util/IdUtil.java
@@ -37,6 +37,26 @@ public class IdUtil {
 	public static String simpleUUID() {
 		return UUID.randomUUID().toString(true);
 	}
+	
+	/**
+	 * 获取随机UUID，使用性能更好的ThreadLocalRandom生成UUID
+	 * 
+	 * @return 随机UUID
+	 * @since 4.1.19
+	 */
+	public static String fastUUID() {
+		return UUID.fastUUID().toString();
+	}
+	
+	/**
+	 * 简化的UUID，去掉了横线，使用性能更好的ThreadLocalRandom生成UUID
+	 * 
+	 * @return 简化的UUID，去掉了横线
+	 * @since 4.1.19
+	 */
+	public static String fastSimpleUUID() {
+		return UUID.fastUUID().toString(true);
+	}
 
 	/**
 	 * 创建MongoDB ID生成策略实现<br>